Updated kit-sortedarray to be compatible with sxe-sortedarray
